# src/main/java/com/anamariafelix/matematico/dto/PersonDTO.java
package com.anamariafelix.matematico.dto;

import jakarta.persistence.Column;
import lombok.*;

@AllArgsConstructor
@NoArgsConstructor
@Getter @Setter @EqualsAndHashCode
public class PersonDTO {

    private Long id;
    private String firstName;
    private String lastName;
    private String address;
    private String gender;
    private Boolean enabled;
}